gcp/functions/cpvton/requirements.txt
Flask==1.0.2
google-cloud-storage==1.16.1
google-cloud-logging==1.15.0

# LIP_JPPNet
matplotlib==3.2.1
numpy==1.18.2    
opencv-python==4.0.0.21  
Pillow==6.0.0     
scipy==1.4.1     
tensorflow==1.14.0
imageio

# cp-vton

